Update src/alert/alert.js
Allow the absence of an explicit alert type to use the default
(yellow) bootstrap color for alerts. I'm using `default` here,
which causes the CSS class `alert-default` to be applied, which
does not exist. Perhaps `unspecified` or `unknown` would be better.
Or we could rewrite `template/alert/alert.html` and `src/alert/alert.js`
so no style is applied when the type is unspecified.